Extract word from dictionary
In secret_file.rb, i have implemented a logic to
extract a word of length between 5 and 12. This
is the secret code.

I was able to do this by getting a array of all
the words using readlines() method and using
sample() method for randomness from the array.
I had to slice the word as it came with newline
character(\n).