Created a game.rb file for the game loop
game.rb file has the Class Game where when
initialized it takes player name as argument
and also calls extract_word function for
secret_word and stores it in its own secret_word
variable. With this i can create a game board
with '_' having the same length as secret word.

there is also a function called game_start that
calls in a welcome message from Display module
and tells the player the secret is generated and
they have 11 turns.