Create a check_choice function with choice as arg
check_choice(choice) first checks for if the
letter is already guessed and shows a message for
it. if the letter is not guessed it is added
to the guessed_letters array and then checks
if it is included in secret_word. If the letter
exists in the secret_word all instances of that
letter are then put up on the board. If the letter
does not exist then incorrect_letter message is
shown and guesses counter is decremented.